+++ glue/src/main/java/org/somda/sdc/glue/provider/localization/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Implementation of localization service provider.
+ */
+@ParametersAreNonnullByDefault
+package org.somda.sdc.glue.provider.localization;
+
+import javax.annotation.ParametersAreNonnullByDefault;