+++ glue/src/main/java/org/somda/sdc/glue/provider/services/factory/ServicesFactory.java
@@ -2,7 +2,12 @@ package org.somda.sdc.glue.provider.services.factory;
 
 import com.google.inject.assistedinject.Assisted;
 import org.somda.sdc.biceps.provider.access.LocalMdibAccess;
+import org.somda.sdc.glue.provider.localization.LocalizationService;
+import org.somda.sdc.glue.provider.localization.LocalizationStorage;
 import org.somda.sdc.glue.provider.services.HighPriorityServices;
+import org.somda.sdc.glue.provider.services.LowPriorityServices;
+
+import javax.annotation.Nullable;
 
 /**
  * Factory to create {@linkplain ServicesFactory} instances.
@@ -15,4 +20,15 @@ public interface ServicesFactory {
      * @return a new instance.
      */
     HighPriorityServices createHighPriorityServices(@Assisted LocalMdibAccess mdibAccess);
+
+    /**
+     * Creates a new {@linkplain LowPriorityServices} instance.
+     *
+     * @param mdibAccess the {@link LocalMdibAccess} instance that is used for response MDIB data.
+     * @param localizationStorage optional localization storage, if provided used by {@link LocalizationService}.
+     *
+     * @return a new instance.
+     */
+    LowPriorityServices createLowPriorityServices(@Assisted LocalMdibAccess mdibAccess,
+                                                  @Assisted @Nullable LocalizationStorage localizationStorage);
 }